deleteStudent 연결만 끊기, 삭제하기 선택할 수 있는 기능 추가

+++ src/student/deleteStudent/deleteStudent.resolvers.js
@@ -4,10 +4,14 @@ import { protectedMutationResovler } from "../../user/user.utils";
 
 export default {
   Mutation: {
-    deleteStudent: protectedMutationResovler(async (_, { teacherEmail, studentId }, { loggedInUser }) => {
-      await Student.deleteOne({ _id: studentId });
-      await StudentList.updateMany({ $pull: { studentId } });
-
+    deleteStudent: protectedMutationResovler(async (_, { disconnectOnly, teacherEmail, studentId, listId }, { loggedInUser }) => {
+      if (disconnectOnly) {
+        await Student.updateOne({ _id: studentId }, { $pull: { listId } });
+        await StudentList.updateOne({ _id: listId }, { $pull: { studentId } });
+      } else {
+        await Student.deleteOne({ _id: studentId });
+        await StudentList.updateMany({ $pull: { studentId } });
+      }
       return { ok: true };
     }),
   },
